Updates chunk invoiced field on invoice create/delete. Missing changes on update.

+++ libs/entity/src/lib/invoice/invoice.resolver.ts
@@ -1,22 +1,27 @@
 import {
-	Resolver,
-	Query,
-	Mutation,
 	Args,
-	Parent,
-	ResolveField,
+	Mutation,
+	Query,
+	Resolver
 } from '@nestjs/graphql';
+import { MongooseError } from 'mongoose';
+import { ChunkRepository } from '../chunk/chunk.repository';
+import { InvoiceCreate, InvoiceEntity, InvoiceUpdate } from './invoice.entity';
 import { InvoiceRepository } from './invoice.repository';
-import { InvoiceEntity, InvoiceCreate, InvoiceUpdate } from './invoice.entity';
 
 @Resolver(() => InvoiceEntity)
 export class InvoiceResolver {
-	constructor(private readonly invoiceRepository: InvoiceRepository) {}
+	constructor(
+		private readonly invoiceRepository: InvoiceRepository,
+		private readonly chunkRepository: ChunkRepository
+	) { }
 
 	@Mutation(() => InvoiceEntity)
+	// @Transaction
 	createInvoice(
 		@Args('createInvoiceInput') createInvoiceInput: InvoiceCreate
 	) {
+		this.chunkRepository.markInvoiced(createInvoiceInput.lines.flatMap((l) => l.chunkIds), true);
 		return this.invoiceRepository.create(createInvoiceInput);
 	}
 
@@ -34,6 +39,7 @@ export class InvoiceResolver {
 	updateInvoice(
 		@Args('updateInvoiceInput') updateInvoiceInput: InvoiceUpdate
 	) {
+		// TODO: Update chunk invoiced field.
 		return this.invoiceRepository.update(
 			updateInvoiceInput._id,
 			updateInvoiceInput
@@ -41,7 +47,12 @@ export class InvoiceResolver {
 	}
 
 	@Mutation(() => InvoiceEntity)
-	removeInvoice(@Args('id', { type: () => String }) id: string) {
+	async removeInvoice(@Args('id', { type: () => String }) id: string) {
+		const invoice = await this.findOne(id);
+		if (!invoice) {
+			throw new MongooseError('Invoice does not exist');
+		}
+		this.chunkRepository.markInvoiced(invoice.lines.flatMap((l) => l.chunkIds), false);
 		return this.invoiceRepository.remove(id);
 	}
 }
